First attempt at RunResultRecorder decopule
